Clean: extract storybook content markups
